# docs/swagger.yaml
definitions:
  app.EmptyObject:
    type: object
  app.EmptySuccessResponse:
    properties:
      code:
        example: 0
        type: integer
      data:
        $ref: '#/definitions/app.EmptyObject'
      msg:
        $ref: '#/definitions/app.EmptyObject'
    type: object
  app.ErrorResponse:
    properties:
      code:
        type: integer
      details:
        items:
          type: string
        type: array
      msg:
        type: string
    type: object
  errcode.Error:
    type: object
  service.CreateAdminRequest:
    properties:
      name:
        example: admin
        maxLength: 20
        minLength: 2
        type: string
      password:
        example: password
        maxLength: 20
        minLength: 5
        type: string
      username:
        example: admin
        maxLength: 20
        minLength: 5
        type: string
    required:
    - name
    - password
    - username
    type: object
  service.LoginAdminRequest:
    properties:
      password:
        maxLength: 20
        minLength: 5
        type: string
      username:
        maxLength: 20
        minLength: 5
        type: string
    required:
    - password
    - username
    type: object
info:
  contact: {}
paths:
  /admins:
    post:
      consumes:
      - application/json
      parameters:
      - description: 新增後台管理員參數
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.CreateAdminRequest'
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/app.EmptySuccessResponse'
        "400":
          description: 請求失敗
          schema:
            $ref: '#/definitions/app.ErrorResponse'
        "409":
          description: 資料已存在
          schema:
            $ref: '#/definitions/app.ErrorResponse'
        "500":
          description: 伺服器異常
          schema:
            $ref: '#/definitions/app.ErrorResponse'
      summary: '[AM-003] 建立後台管理員'
      tags:
      - Admin
  /admins/auth/login:
    post:
      parameters:
      - description: 登入請求參數
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.LoginAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/app.EmptySuccessResponse'
        "400":
          description: 請求失敗
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 伺服器異常
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: '[AM-001] 登入後台管理員'
      tags:
      - Admin
  /admins/auth/logout:
    post:
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/app.EmptySuccessResponse'
        "400":
          description: 請求失敗
          schema:
            $ref: '#/definitions/app.ErrorResponse'
        "500":
          description: 伺服器異常
          schema:
            $ref: '#/definitions/app.ErrorResponse'
      summary: '[AM-002] 登出後台管理員'
      tags:
      - Admin
  /articles:
    delete:
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/app.EmptySuccessResponse'
        "400":
          description: 請求失敗
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 伺服器異常
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: '[AC-002] 取得文章列表'
      tags:
      - Aritcle
    post:
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/app.EmptySuccessResponse'
        "400":
          description: 請求失敗
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 伺服器異常
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: '[AC-003] 新增文章'
      tags:
      - Aritcle
  /articles/{id}:
    delete:
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/app.EmptySuccessResponse'
        "400":
          description: 請求失敗
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 伺服器異常
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: '[AC-005] 刪除文章'
      tags:
      - Aritcle
    get:
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/app.EmptySuccessResponse'
        "400":
          description: 請求失敗
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 伺服器異常
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: '[AC-001] 取得文章資訊'
      tags:
      - Aritcle
    put:
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/app.EmptySuccessResponse'
        "400":
          description: 請求失敗
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 伺服器異常
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: '[AC-004] 更新文章'
      tags:
      - Aritcle
  /artworks:
    delete:
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/app.EmptySuccessResponse'
        "400":
          description: 請求失敗
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 伺服器異常
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: '[AC-002] 取得作品列表'
      tags:
      - Artwork
    post:
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/app.EmptySuccessResponse'
        "400":
          description: 請求失敗
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 伺服器異常
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: '[AC-003] 新增作品'
      tags:
      - Artwork
  /artworks/{id}:
    delete:
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/app.EmptySuccessResponse'
        "400":
          description: 請求失敗
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 伺服器異常
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: '[AC-005] 刪除作品'
      tags:
      - Artwork
    get:
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/app.EmptySuccessResponse'
        "400":
          description: 請求失敗
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 伺服器異常
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: '[AC-001] 取得作品資訊'
      tags:
      - Artwork
    put:
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/app.EmptySuccessResponse'
        "400":
          description: 請求失敗
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 伺服器異常
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: '[AC-004] 更新作品'
      tags:
      - Artwork
  /users:
    get:
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/app.EmptySuccessResponse'
        "400":
          description: 請求失敗
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 伺服器異常
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: '[US-005] 取得使用者列表'
      tags:
      - User
  /users/{id}:
    delete:
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/app.EmptySuccessResponse'
        "400":
          description: 請求失敗
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 伺服器異常
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: '[US-007] 刪除使用者'
      tags:
      - User
    get:
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/app.EmptySuccessResponse'
        "400":
          description: 請求失敗
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 伺服器異常
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: '[US-004] 取得使用者資訊'
      tags:
      - User
    put:
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/app.EmptySuccessResponse'
        "400":
          description: 請求失敗
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 伺服器異常
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: '[US-006] 更新使用者資訊'
      tags:
      - User
  /users/auth/login:
    post:
      parameters:
      - description: 登入請求參數
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.LoginAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/app.EmptySuccessResponse'
        "400":
          description: 請求失敗
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 伺服器異常
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: '[US-001] 登入使用者'
      tags:
      - User
  /users/auth/logout:
    post:
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/app.EmptySuccessResponse'
        "400":
          description: 請求失敗
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 伺服器異常
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: '[US-002] 登出使用者'
      tags:
      - User
  /users/auth/signup:
    post:
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/app.EmptySuccessResponse'
        "400":
          description: 請求失敗
          schema:
            $ref: '#/definitions/errcode.Error'
        "409":
          description: 資料已存在
          schema:
            $ref: '#/definitions/app.ErrorResponse'
        "500":
          description: 伺服器異常
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: '[US-003] 註冊帳號'
      tags:
      - User
swagger: "2.0"